perf(client): cut initial load payload and unblock first render
Profiling repeated reloads showed first meaningful paint at ~1.4-1.5s with
the main thread busy 1501ms out of 2200ms. Four causes, all addressed here.

Restore code splitting. pages/User/User.jsx was both a lazy route and the
module 48 other files imported Avatar/NameWithAvatar/MediaEmbed from, so the
whole User page tree landed in the initial bundle. Split the shared widgets
into UserWidgets.jsx and leave User.jsx as just the route. Miniprofile's
import of three icon constants from the 2639-line Profile.jsx re-created the
same cascade, so those move to profileIcons.js. Miniprofile (pulls d3) and
Auth (pulls firebase/auth) only render inside an opened popover/dialog, so
both are now lazy. Eager module graph: 158 modules -> 64.

Stop inlining images into the render-blocking stylesheet. roles.css alone
references ~780 role icons, all base64-inlined at rsbuild's default 4KB
threshold, making index.css 906KB of which 87% was data: URIs.

Make the auto-contrast lookup lazy. generateContrastLookup eagerly filled
11 x 360 cells -- 39,600 Color.contrast() calls, ~150ms of synchronous work
per load, landing between "user data arrived" and "routes render". Cells are
now solved on first use and cached; output is unchanged (verified identical
across 59,400 cells over 15 background/threshold combinations). Since it is
cheap now it is derived with useMemo rather than an effect, which also means
the first paint has adjusted colors instead of flashing unadjusted ones.

Parallelize /api/user/info. Its four independent lookups ran as sequential
round trips; this endpoint gates the client's first render.

Initial render-blocking payload: 2840KB -> 961KB raw, 1092KB -> 284KB gzip.
Measured on repeated reloads: loadEventStart 775/795ms -> 483/501ms, FMP
1417/1511ms -> 1209/1332ms.
