+++ react_main/src/components/EmoteReactions.jsx
@@ -4,7 +4,7 @@ import update from "immutability-helper";
 
 import { UserContext } from "Contexts";
 import { useErrorAlert } from "components/Alerts";
-import { NameWithAvatar } from "pages/User/User";
+import { NameWithAvatar } from "pages/User/UserWidgets";
 import EmotePicker from "components/EmotePicker";
 import { Emote, CustomEmote, Emotes } from "components/Emotes";
 import { Box, Popover, Stack, Typography } from "@mui/material";
