+++ react_main/src/components/Popover.jsx
@@ -11,12 +11,12 @@ import {
   getAlignmentColor,
 } from "components/Setup";
 import { useErrorAlert } from "components/Alerts";
-import { NameWithAvatar } from "pages/User/User";
+import { NameWithAvatar } from "pages/User/UserWidgets";
 
 import { Box, Divider, Link, Popover, Stack, Typography } from "@mui/material";
 import { usePopoverOpen } from "hooks/usePopoverOpen";
 import { GameSettingCount } from "./Roles";
-import { KUDOS_ICON } from "pages/User/Profile";
+import { KUDOS_ICON } from "pages/User/profileIcons";
 
 export function PopoverContent({ title, content, page = null, icon = <></>, subTitle = <></> }) {
   let wrappedTitle = (
