+++ react_main/src/components/RecentForumReplies.jsx
@@ -3,7 +3,7 @@ import axios from "axios";
 import { Link } from "react-router-dom";
 
 import { Paper, Stack, Typography, Box } from "@mui/material";
-import { Avatar } from "pages/User/User";
+import { Avatar } from "pages/User/UserWidgets";
 import { Time } from "components/Basic";
 import LobbySidebarPanel from "pages/Play/LobbyBrowser/LobbySidebarPanel";
