+++ react_main/src/components/Strategies.jsx
@@ -25,7 +25,7 @@ import { useErrorAlert } from "components/Alerts";
 import { TextEditor } from "components/Form";
 import CustomMarkdown from "components/CustomMarkdown";
 import { VoteWidget } from "components/VoteWidget";
-import { NameWithAvatar } from "pages/User/User";
+import { NameWithAvatar } from "pages/User/UserWidgets";
 import { Loading } from "components/Loading";
 import surprisedFace from "images/emotes/surprised.webp";
 import sadFace from "images/emotes/sad.webp";
