+++ react_main/src/pages/Fame/Donors.jsx
@@ -15,7 +15,7 @@ import {
 } from "@mui/material";
 import { useErrorAlert } from "../../components/Alerts";
 import { UserContext, SiteInfoContext } from "../../Contexts";
-import { Avatar } from "../User/User";
+import { Avatar } from "../User/UserWidgets";
 import { TextEditor, UserSearchSelect } from "../../components/Form";
 import CustomMarkdown from "../../components/CustomMarkdown";
