+++ react_main/src/pages/Game/ChessGame.jsx
@@ -14,7 +14,7 @@ import {
   GameTypeContext,
 } from "./Game";
 import { GameContext } from "../../Contexts";
-import { Avatar } from "../User/User";
+import { Avatar } from "../User/UserWidgets";
 import { SideMenu } from "./Game";
 import { useIsPhoneDevice } from "hooks/useIsPhoneDevice";
