+++ react_main/src/pages/User/profileIcons.js
@@ -1,0 +1,10 @@
+// Profile stat icons. Kept in their own leaf module so that components which
+// only need the icons (Miniprofile, Popover) don't pull in the whole Profile
+// page tree — and with it Comments, EmotePicker, ModerationSideDrawer, etc.
+export const KUDOS_ICON = require(`images/kudos.png`);
+export const KARMA_ICON = require(`images/karma.png`);
+export const POINTS_ICON = require(`images/points.png`);
+export const POINTS_NEGATIVE_ICON = require(`images/pointsNegative.png`);
+export const PRESTIGE_ICON = require(`images/prestige.png`);
+export const ACHIEVEMENTS_ICON = require(`images/achievements.png`);
+export const DAILY_ICON = require(`images/dailyChallenges.png`);
