+++ react_main/src/utilsFolder/autoContrast.jsx
@@ -9,49 +9,66 @@ import { useContext } from 'react';
 const DEFAULT_MINIMUM_CONTRAST = 3.5; // WCAG 2.1
 const CHROMA_BANDS = 10;
 
-export function generateContrastLookup(backgroundColor, minimumContrast = DEFAULT_MINIMUM_CONTRAST) {
-  const target = Color(backgroundColor);
-  const darkMode = target.isDark();
+// Solves for the lowest (dark mode: highest) lightness at the given chroma band
+// and hue that still meets `minimumContrast` against `target`.
+// Binary search with 10 iterations — same search as before, one cell at a time.
+function solveBandHue(target, darkMode, minimumContrast, chromaBand, hue) {
+  const chroma = (chromaBand * 145) / CHROMA_BANDS;
 
-  if (minimumContrast === 0) { 
-    return null;
-  }
-
-  let contrastLookup = {};
-  for (let c = 0; c < CHROMA_BANDS + 1; c++) {
-    contrastLookup[c] = {};
-    const chroma = c * 145 / CHROMA_BANDS;
-
-    for (let h = 0; h < 360; h++) {
-      // Binary search with 10 iterations for the lowest lightness that meets contrast
-      let low = 0, high = 100;
-      let bestLightness = 0;
-      for (let i = 0; i < 10; i++) {
-        const mid = (low + high) / 2;
-        const currentRatio = target.contrast(Color.lch(mid, chroma, h));
-        if (currentRatio >= minimumContrast) {
-          bestLightness = mid;
-          if (darkMode) {
-            high = mid;
-          }
-          else {
-            low = mid
-          }
-        } else {
-          if (darkMode) {
-            low = mid
-          }
-          else {
-            high = mid;
-          }
-        }
+  let low = 0, high = 100;
+  let bestLightness = 0;
+  for (let i = 0; i < 10; i++) {
+    const mid = (low + high) / 2;
+    const currentRatio = target.contrast(Color.lch(mid, chroma, hue));
+    if (currentRatio >= minimumContrast) {
+      bestLightness = mid;
+      if (darkMode) {
+        high = mid;
       }
-
-      contrastLookup[c][h] = Color.lch(bestLightness, chroma, h).hex();
+      else {
+        low = mid
+      }
+    } else {
+      if (darkMode) {
+        low = mid
+      }
+      else {
+        high = mid;
+      }
     }
   }
 
-  return contrastLookup;
+  return Color.lch(bestLightness, chroma, hue).hex();
+}
+
+// Returns a memoizing lookup rather than a precomputed table.
+//
+// This used to eagerly fill all (CHROMA_BANDS + 1) x 360 cells, which is 39,600
+// Color.contrast() calls — ~150ms of synchronous main-thread work on every page
+// load, landing right between "user data arrived" and "routes render". A page
+// only ever asks for the handful of (band, hue) pairs it actually paints, so
+// cells are now solved on first use and cached. Results are identical to the
+// eager table; only when the work happens has changed.
+export function generateContrastLookup(backgroundColor, minimumContrast = DEFAULT_MINIMUM_CONTRAST) {
+  if (minimumContrast === 0) {
+    return null;
+  }
+
+  const target = Color(backgroundColor);
+  const darkMode = target.isDark();
+  const cache = new Map();
+
+  return {
+    get(chromaBand, hue) {
+      const key = chromaBand * 360 + hue;
+      let hex = cache.get(key);
+      if (hex === undefined) {
+        hex = solveBandHue(target, darkMode, minimumContrast, chromaBand, hue);
+        cache.set(key, hex);
+      }
+      return hex;
+    },
+  };
 }
 
 export function autoContrastColor(sourceColor, backgroundColor, contrastLookup, minimumContrast = DEFAULT_MINIMUM_CONTRAST) {
@@ -76,7 +93,7 @@ export function autoContrastColor(sourceColor, backgroundColor, contrastLookup, 
     const colorLch = color.lch().array();
     const chromaBand = Math.min(CHROMA_BANDS, Math.trunc(colorLch[1] / CHROMA_BANDS));
     const hue = Math.trunc(colorLch[2]);
-    return contrastLookup[chromaBand][hue];
+    return contrastLookup.get(chromaBand, hue);
   }
 }
